Refactor the component that shows movie list
